Move App\Core\Controller to App\Core\LegacyController and improve it
