+++ app/Core/BackendController.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Controller - A base Controller for the included example Modules.
+ * BackendController - A backend Controller for the included example Modules.
  *
  * @author Virgil-Adrian Teaca - virgil@giulianaeassociati.com
  * @version 3.0
