+++ system/Core/Controller.php
@@ -9,10 +9,7 @@
 namespace Core;
 
 use Core\Config;
-use Core\Language;
-use Http\Response;
 use Routing\Controller as BaseController;
-use Routing\Route;
 use Support\Contracts\RenderableInterface as Renderable;
 use Template\Template as Layout;
 
@@ -41,13 +38,6 @@ abstract class Controller extends BaseController
      */
     protected $layout = 'default';
 
-    /**
-     * Language variable to use the languages class.
-     *
-     * @var string
-     */
-    public $language = null;
-
 
     /**
      * On the initial run, create an instance of the config class and the view class.
@@ -57,11 +47,6 @@ abstract class Controller extends BaseController
         // Setup the used Template to default, if it is not already defined.
         if (($this->layout !== false) && ! isset($this->template)) {
             $this->template = Config::get('app.template');
-        }
-
-        // Initialise the Language object.
-        if ($this->language !== false) {
-            $this->language = Language::getInstance();
         }
     }
 
@@ -79,41 +64,11 @@ abstract class Controller extends BaseController
             // we will assume we want to render it using the Controller's templated environment.
 
             if (($this->layout !== false) && (! $response instanceof Layout)) {
-                return Template::make($this->layout, $this->template)->with('content', $response);
+                $response = Template::make($this->layout, $this->template)->with('content', $response);
             }
-        } else if (is_null($response)) {
-            // If the response which is returned from the Controller's Action is null and we have
-            // View instances on View's Legacy support, we will assume that we are on Legacy Mode.
-
-            $items = View::getItems();
-
-            $headers = View::getHeaders();
-
-            // Render the View instances to response.
-            $response = '';
-
-            foreach ($items as $item) {
-                $response .= $item->render();
-            }
-
-            // Create a Response instance and return it.
-            return new Response($response, 200, $headers);
         }
 
-        return $response;
-    }
-
-    /**
-     * Return a translated string.
-     * @return string
-     */
-    protected function trans($str, $code = LANGUAGE_CODE)
-    {
-        if ($this->language instanceof Language) {
-            return $this->language->get($str, $code);
-        }
-
-        return $str;
+        return parent::processResponse($response);
     }
 
     /**
